move gamecontext
